requirements.txt
asgiref==3.5.2
chardet==5.0.0
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
lml==0.1.0
mysqlclient==2.1.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
